[FIX] небольшой рефакторинг

+++ app/src/main/java/com/example/numbersfact/api/ApiJSON.kt
@@ -2,7 +2,7 @@ package com.example.numbersfact.api
 
 data class ApiJSON(
     val text:String,
-    val number:Long,
+    val number:Double, // Сюда прилетают числа больше, чем Long
     val found:Boolean,
     val type:String
 )
